refactor(ui): consolidate ISidebarMethodOptions interface and remove unused files
