--- packages/ui/src/views/components/sidebar/interface.ts
@@ -1,33 +1,0 @@
-/**
- * Copyright 2023-present DreamNum Co., Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { CSSProperties } from 'react';
-import type { ICustomLabelProps } from '../../../components/custom-label/CustomLabel';
-
-export interface ISidebarMethodOptions {
-    id?: string;
-    header?: ICustomLabelProps;
-    children?: ICustomLabelProps;
-    bodyStyle?: CSSProperties;
-    footer?: ICustomLabelProps;
-
-    visible?: boolean;
-
-    width?: number | string;
-
-    onClose?: () => void;
-    onOpen?: () => void;
-}
